TransformのLockAt関数を修正。 テキストのずれを修正。 CameraのWolrdToScreen関数を改善。 SetActive関数で子供たちにも同時に設定出来るようになった。 GameObjectManagerで優先度がマイナスにならないように修正。 Supportに文字列置き換え関数のStrReplace追加。 SupportのLoadCSVDataの使用を変更。

+++ game/fbEngine/fbEngine/_Object/_GameObject/FPS.cpp
@@ -6,7 +6,7 @@ void FPS::Start()
 	
 	_Text->SetSize(40.0f);
 	_Text->SetString(L"FPS:0");
-	_Text->SetFormat(fbText::TextFormatE::LEFT);
+	_Text->SetFormat((int)fbText::TextFormatE::LEFT);
 }
 
 void FPS::Update()
